inventory-mcp/requirements.txt
fastmcp>=2.10.0
# [FIX] Pin below 4.0.0 to avoid namespace namespace conflicts
google-cloud-asset>=3.20.0,<4.0.0
# [FIX] Remove core pin to let the library resolve its own dependencies
google-cloud-core>=2.3.0
proto-plus>=1.22.0
protobuf>=4.25.0
python-dotenv>=1.0.0
starlette>=0.36.0
uvicorn[standard]>=0.27.0
